fix error

+++ FMS.BusinessObject/Dto/EpafDto.cs
@@ -8,7 +8,7 @@ namespace FMS.BusinessObject.Dto
 {
     public class EpafDto
     {
-        public int MstEpafId { get; set; }
+        public long MstEpafId { get; set; }
         public int DocumentType { get; set; }
         public string EpafAction { get; set; }
         public string EmployeeId { get; set; }
